Floppy uses midi input

+++ floppy-music-macros/src/lib.rs
@@ -27,7 +27,7 @@ pub fn midi_note_periods(input: TokenStream) -> TokenStream {
         .into_iter()
         .map(|note| ast.a4_freq as f32 * 2.0f32.powf((note as f32 - ast.a4_midi as f32) / 12.0))
         .map(|freq| 1.0 / freq * 1_000_000.0)
-        .map(|period| period as u64)
+        .map(|period| period as u32)
         .collect::<Vec<_>>();
 
     let expanded = quote! {
